Initial files (#1)
* Create CMakeLists.txt

main `cmake` configuration file. Not finished, you need to update this before using.

* Create src/CMakeLists.txt

`cmake` configuration file for source files (code itself). Not finished, you need to update this before using.

* Create tests/CMakeLists.txt

`cmake` configuration file for test files (code itself). Not finished, you need to update this before using.

* Create fruit.f90

Fortran test framework, fruit. Use it as it is. Only change it if really needed.

* Create fruit.f90

Fortran test framework utils, fruit. Use it as it is. Only change it if really needed.

* Create test_io.f90

Example tests for class_io. More can/should be added.

* Create test_mesh.f90

Example tests for `class_mesh.f90`. More can/should be added.

* Create class_mesh.f90

It is not complete, you can modify the type as well. Rewrite it according to your needs. And rewrite the tests if needed.

* Create class_io_netcdf4.f90

It is very close to working code, just clean the MPI part as you do not need it. 

If you will not going to use netcdf4 this just an example.

* Delete README.md

I do not think we need this here.   Add a section in the main README.md named **Contributing**.

* Create precision.f90

It works, no need to modify. Make it sure all your code relies on this variables `ik` and `rk`.

* Update README.md

Add Compilation and Contributing sections.

Add figure of potential result.

* Create TargetDistclean.cmake

Additional cmake tool/command.